Modify extension
-Change from js to jsx
