--- reactExamples.js
@@ -1,23 +1,0 @@
-//Conditionals
-
-var React = require('react');
-var ReactDOM = require('react-dom');
-
-if (user.age >= drinkingAge) {
-  var message = (
-    <h1>
-      Hey, check out this alcoholic beverage!
-    </h1>
-  );
-} else {
-  var message = (
-    <h1>
-      Hey, check out these earrings I got at Claire's!
-    </h1>
-  );
-}
-
-ReactDOM.render(
-  message,
-  document.getElementById('app')
-);
